refactor: Content interfaces into their own sub-package.
