+++ app/src/main/java/no/usn/mob3000/data/repository/content/NewsRepository.kt
@@ -6,7 +6,7 @@ import no.usn.mob3000.data.source.remote.auth.AuthDataSource
 import no.usn.mob3000.data.source.remote.content.NewsDataSource
 import no.usn.mob3000.domain.model.content.NewsData
 import no.usn.mob3000.domain.model.content.NewsUpdateData
-import no.usn.mob3000.domain.repository.INewsRepository
+import no.usn.mob3000.domain.repository.content.INewsRepository
 
 /**
  * Repository class responsible for managing news-related operations.
