+++ app/src/main/java/no/usn/mob3000/domain/usecase/content/docs/DeleteDocsUseCase.kt
@@ -1,7 +1,7 @@
 package no.usn.mob3000.domain.usecase.content.docs
 
 import no.usn.mob3000.data.repository.content.DocsRepository
-import no.usn.mob3000.domain.repository.IDocsRepository
+import no.usn.mob3000.domain.repository.content.IDocsRepository
 
 /**
  * Use case for deleting news.
