+++ app/src/main/java/no/usn/mob3000/domain/usecase/content/docs/InsertDocsUseCase.kt
@@ -1,6 +1,6 @@
 package no.usn.mob3000.domain.usecase.content.docs
 
-import no.usn.mob3000.domain.repository.IDocsRepository
+import no.usn.mob3000.domain.repository.content.IDocsRepository
 
 /**
  * Use case for inserting docs.
