+++ app/src/main/java/no/usn/mob3000/domain/usecase/content/faq/DeleteFAQUseCase.kt
@@ -1,7 +1,7 @@
 package no.usn.mob3000.domain.usecase.content.faq
 
 import no.usn.mob3000.data.repository.content.FAQRepository
-import no.usn.mob3000.domain.repository.IFAQRepository
+import no.usn.mob3000.domain.repository.content.IFAQRepository
 
 /**
  * Use case for deleting FAQ.
