+++ app/src/main/java/no/usn/mob3000/domain/usecase/content/news/DeleteNewsUseCase.kt
@@ -1,7 +1,7 @@
 package no.usn.mob3000.domain.usecase.content.news
 
 import no.usn.mob3000.data.repository.content.NewsRepository
-import no.usn.mob3000.domain.repository.INewsRepository
+import no.usn.mob3000.domain.repository.content.INewsRepository
 
 /**
  * Use case for deleting news.
